+++ assets/javascripts/kitten/components/cards/reward-card/starred-badge.js
@@ -67,5 +67,4 @@ const styles = {
   },
 }
 
-
 export const RewardCardStarredBadge = Radium(RewardCardStarredBadgeBase)