+++ assets/javascripts/kitten/components/cards/team-card.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react'
+import { TeamCardTitle } from 'kitten/components/cards/team-card/title'
+import { TeamCardImage } from 'kitten/components/cards/team-card/image'
+import { TeamCardIcons } from 'kitten/components/cards/team-card/icons'
+
+export class TeamCard extends Component {
+  static Image = TeamCardImage
+  static Title = TeamCardTitle
+  static Icons = TeamCardIcons
+
+  render = () => this.props.children
+}