+++ assets/javascripts/kitten/helpers/utils/deprecated.js
@@ -3,8 +3,11 @@ import React, { Component } from 'react'
 export class Deprecated extends Component {
   componentWillMount() {
     if (process.env.NODE_ENV === 'development') {
-      const componentName = React.Children.only(this.props.children)._owner.type
-        .name
+      const owner = React.Children.only(this.props.children)._owner
+
+      if (!owner) return
+
+      const componentName = owner.type.name
 
       console.warn(
         `Warning: ${componentName} is deprecated.`,